pyproject.toml
[tool.poetry]
name = "devblog"
version = "0.1.0"
description = "My devblog"
authors = ["Olivier Philippon <olivier@dunsap.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs-material = "^8.3.9"
mkdocs-awesome-pages-plugin = "^2.7.0"
# @link https://github.com/aklajnert/mkdocs-simple-hooks
mkdocs-simple-hooks = "^0.1.5"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[tool.black]
line-length = 119
target-version = ['py310']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
